% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cppTemplate-package}
\alias{cppTemplate-package}
\title{Dummy package to test out C++}
\description{
Some boring description here
}
\author{
Me \email{me@gmail.com}

Maintainer: None
}
\keyword{package}
